migrated sqldb wizard to generic ResourceTypeWizard
