+++ extensions/resource-deployment/src/ui/deployAzureSQLDBWizard/deployAzureSQLDBWizardModel.ts
@@ -5,9 +5,22 @@
 
 import { EOL } from 'os';
 import * as azdata from 'azdata';
-import { Model } from '../model';
+import * as vscode from 'vscode';
+import * as constants from './constants';
+import axios, { AxiosRequestConfig } from 'axios';
+import { ResourceTypeModel } from '../resourceTypeModel';
+import { INotebookService } from '../../services/notebookService';
+import { IToolsService } from '../../services/toolsService';
+import { AzureSQLDBDeploymentProvider } from '../../interfaces';
+import { ResourceTypeWizard } from '../resourceTypeWizard';
+import { ResourceTypePage } from '../resourceTypePage';
+import { DatabaseSettingsPage } from './pages/databaseSettingsPage';
+import { AzureSQLDBSummaryPage } from './pages/summaryPage';
+import { AzureSettingsPage } from './pages/azureSettingsPage';
 
-export class DeployAzureSQLDBWizardModel extends Model {
+export class DeployAzureSQLDBWizardModel extends ResourceTypeModel {
+	private cache: Map<string, any> = new Map();
+
 	public azureAccount!: azdata.Account;
 	public securityToken!: any;
 	public azureSubscription!: string;
@@ -28,9 +41,149 @@ export class DeployAzureSQLDBWizardModel extends Model {
 	public firewallRuleName!: string;
 	public databaseCollation!: string;
 
+	public get notebookService(): INotebookService {
+		return this.wizard.notebookService;
+	}
 
-	constructor() {
-		super();
+	public get toolService(): IToolsService {
+		return this.wizard.toolsService;
+	}
+
+	constructor(public sqldbProvider: AzureSQLDBDeploymentProvider, wizard: ResourceTypeWizard) {
+		super(sqldbProvider, wizard);
+		this.wizard.wizardObject.title = constants.WizardTitle;
+	}
+
+	initialize(): void {
+		this.wizard.setPages(this.getPages());
+		this.wizard.wizardObject.generateScriptButton.hidden = true;
+		this.wizard.wizardObject.doneButton.label = constants.WizardDoneButtonLabel;
+	}
+
+	async onOk(): Promise<void> {
+		await this.scriptToNotebook();
+	}
+
+	onCancel(): void {
+	}
+
+	private getPages(): ResourceTypePage[] {
+		const pages: ResourceTypePage[] = [];
+		pages.push(new AzureSettingsPage(this));
+		pages.push(new DatabaseSettingsPage(this));
+		pages.push(new AzureSQLDBSummaryPage(this));
+		return pages;
+	}
+
+	public async getRequest(url: string, useCache = false): Promise<any> {
+		if (useCache) {
+			if (this.cache.has(url)) {
+				return this.cache.get(url);
+			}
+		}
+		let token = this.securityToken.token;
+		const config: AxiosRequestConfig = {
+			headers: {
+				'Content-Type': 'application/json',
+				'Authorization': `Bearer ${token}`
+			},
+			validateStatus: () => true // Never throw
+		};
+		const response = await axios.get(url, config);
+		if (response.status !== 200) {
+			let errorMessage: string[] = [];
+			errorMessage.push(response.status.toString());
+			errorMessage.push(response.statusText);
+			if (response.data && response.data.error) {
+				errorMessage.push(`${response.data.error.code} : ${response.data.error.message}`);
+			}
+			vscode.window.showErrorMessage(errorMessage.join(EOL));
+		}
+		if (useCache) {
+			this.cache.set(url, response);
+		}
+		return response;
+	}
+
+	public createFormRowComponent(view: azdata.ModelView, title: string, description: string, component: azdata.Component, required: boolean): azdata.FlexContainer {
+
+		component.updateProperties({
+			required: required,
+			width: '480px'
+		});
+
+		const labelText = view.modelBuilder.text()
+			.withProperties<azdata.TextComponentProperties>(
+				{
+					value: title,
+					width: '250px',
+					description: description,
+					requiredIndicator: required,
+				})
+			.component();
+
+		labelText.updateCssStyles({
+			'font-weight': '400',
+			'font-size': '13px',
+		});
+
+		const flexContainer = view.modelBuilder.flexContainer()
+			.withLayout(
+				{
+					flexFlow: 'row',
+					alignItems: 'center',
+				})
+			.withItems(
+				[labelText, component],
+				{
+					CSSStyles: { 'margin-right': '5px' }
+				})
+			.component();
+		return flexContainer;
+	}
+
+	public changeComponentDisplay(component: azdata.Component, display: ('none' | 'block')) {
+		component.updateProperties({
+			required: display === 'block'
+		});
+		component.updateCssStyles({
+			display: display
+		});
+	}
+
+	public changeRowDisplay(container: azdata.FlexContainer, display: ('none' | 'block')) {
+		container.items.map((component) => {
+			component.updateProperties({
+				required: (display === 'block'),
+			});
+			component.updateCssStyles({
+				display: display,
+			});
+		});
+	}
+
+	public addDropdownValues(component: azdata.DropDownComponent, values: azdata.CategoryValue[], width?: number) {
+		component.updateProperties({
+			values: values,
+			width: '480px'
+		});
+	}
+
+	public showErrorMessage(message: string) {
+		this.wizard.wizardObject.message = {
+			text: message,
+			level: azdata.window.MessageLevel.Error
+		};
+	}
+
+	private async scriptToNotebook(): Promise<void> {
+		const variableValueStatements = this.getCodeCellContentForNotebook();
+		const insertionPosition = 2; // Cell number 2 is the position where the python variable setting statements need to be inserted in this.wizardInfo.notebook.
+		try {
+			await this.notebookService.openNotebookWithEdits(this.sqldbProvider.azureSQLDBWizard.notebook, variableValueStatements, insertionPosition);
+		} catch (error) {
+			vscode.window.showErrorMessage(error);
+		}
 	}
 
 	public getCodeCellContentForNotebook(): string[] {
