+++ extensions/resource-deployment/src/ui/deployAzureSQLDBWizard/pages/summaryPage.ts
@@ -4,23 +4,23 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as azdata from 'azdata';
-import { WizardPageBase } from '../../wizardPageBase';
-import { DeployAzureSQLDBWizard } from '../deployAzureSQLDBWizard';
 import * as constants from '../constants';
 import { SectionInfo, LabelPosition, FontWeight, FieldType } from '../../../interfaces';
 import { createSection } from '../../modelViewUtils';
+import { BasePage } from '../../deployAzureSQLVMWizard/pages/basePage';
+import { DeployAzureSQLDBWizardModel } from '../deployAzureSQLDBWizardModel';
 
-export class AzureSQLDBSummaryPage extends WizardPageBase<DeployAzureSQLDBWizard> {
+export class AzureSQLDBSummaryPage extends BasePage {
 
 	private formItems: azdata.FormComponent[] = [];
 	private _form!: azdata.FormBuilder;
 	private _view!: azdata.ModelView;
 
-	constructor(wizard: DeployAzureSQLDBWizard) {
+	constructor(private _model: DeployAzureSQLDBWizardModel) {
 		super(
 			'Summary',
 			'',
-			wizard
+			_model.wizard
 		);
 
 	}
@@ -41,7 +41,7 @@ export class AzureSQLDBSummaryPage extends WizardPageBase<DeployAzureSQLDBWizard
 
 		this.formItems = [];
 
-		let model = this.wizard.model;
+		let model = this._model;
 
 		const labelWidth = '150px';
 		const inputWidth = '400px';
